Get top miRNA names in common between two topics of a data frame

Get top miRNA names in common between two topics of a data frame.

get_shared_mir_df(
  df,
  top = 5,
  topic = NULL,
  col.topic = Topic,
  col.mir = miRNA,
  col.pmid = PMID
)

Arguments

df
Data frame containing at least two topics and miRNA names.

top
Integer. Number of top miRNA names to extract for both topics.

topic
String. Vector of strings containing topic names to compare miRNA names for. If topic = NULL, topic defaults to all topic names contained in col.topic in df. topic must only contain two topic names.

col.topic
Symbol. Column containing topic names.

col.mir
Symbol. Column containing miRNA names.

col.pmid
Symbol. Column containing PubMed-IDs.

Value

Character vector containing miRNA names in common between two topics.

Details

Get top miRNA names in common between two topics of a data frame. get_shared_mir_df() compares the top miRNA names of two topics in a data frame and returns the miRNA names in common.

See also