Plot frequency of patient scores in abstracts

Plot frequency of patient scores in abstracts.

plot_score_patients(
  df,
  keywords = patients_keywords,
  case = FALSE,
  bins = NULL,
  colour = "steelblue3",
  col.abstract = Abstract,
  col.pmid = PMID,
  title = NULL
)

Arguments

df
Data frame containing abstracts.

keywords
Character vector. Vector containing keywords. The score is calculated based on these keywords. How much weight a keyword in keywords carries is determined how often it is present in keywords, e.g. if a keyword is mentioned twice in keywords and it is mentioned only once in an abstract, it adds 2 points to the score.

case
Boolean. If case = TRUE, terms contained in keywords are case sensitive. If case = FALSE, terms contained in keywords are case insensitive.

bins
Integer. Specifies how many bins are used to plot the distribution. If bins = NULL, bins are calculated over the whole range of scores, with one bin per score.

colour
String. Colour of histogram.

col.abstract
Symbol. Column containing abstracts.

col.pmid
Symbol. Column containing PubMed-IDs.

title
String. Plot title.

Value

Histogram displaying the distribution of patient scores in abstracts.

Details

Plots a frequency distribution of patient scores in abstracts of a data frame. The patient score is influenced by the choice of terms in keywords. Plotting the distribution can help deciding if the terms are well-chosen, or in choosing the right threshold to decide which abstracts are considered to contain patient material

See also